Added unit tests to ChemReduce.
Added unit tests to ChemReduce. Tests the computation of the coefficients
of the linear error constraint, as well as the right-hand side vector
of those constraints. Floating point comparisons are made with an absolute
tolerance of 1e-6; two tests fail at the 1e-7 level. This absolute tolerance
seems a bit high, so it may be worth investigating it.