Refactored ChemReduce.
Refactored the main routines in ChemReduce (specifically, within
chemReduce.py). Pulled out the generation of coefficients for the
linear error constraints so that the integer linear program has
constraints that look like - b <= A * z <= b, where z is a vector
of binary variables, A is a matrix, and b is a vector (all real
unless otherwise noted). Also refactored out routines to calculate
data needed for constraints, so that they may be reused to set up
unit tests.