Changed authors in license.
Changed authors in license to remove William H. Green and Paul I.
Barton, because they did not contribute to writing the code. (They
did, however, conceive of the ideas for the algorithms, so they are
credited in a new version of the README.) Added MIT to satisfy
copyright requirements for their Technology Licensing Office.

License is still a BSD-3 license.